requirements/base.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0 ; python_version >= '3.6'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0 ; python_version >= '3.7'
djoser==2.1.0
idna==3.4 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
psycopg2==2.9.5
pycparser==2.21
pyjwt==2.6.0 ; python_version >= '3.7'
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.3.0 ; python_version >= '3.6'
sqlparse==0.4.3 ; python_version >= '3.5'
tzdata==2022.7 ; sys_platform == 'win32'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
